Move two firebase DB functions to a source file
