+++ app/src/main/java/alison/fivethingskotlin/ViewModels/FiveThingsViewModel.kt
@@ -1,5 +1,6 @@
 package alison.fivethingskotlin.ViewModels
 
+import alison.fivethingskotlin.Models.FirebaseSource
 import alison.fivethingskotlin.Models.FiveThings
 import alison.fivethingskotlin.Util.getDatabaseStyleDate
 import alison.fivethingskotlin.Util.getDateFromDatabaseStyle
@@ -12,7 +13,6 @@ import android.util.Log
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.database.*
 import java.util.*
-import kotlin.collections.ArrayList
 
 
 class FiveThingsViewModel(private val user: FirebaseUser): ViewModel() {
@@ -20,37 +20,10 @@ class FiveThingsViewModel(private val user: FirebaseUser): ViewModel() {
     private var database = FirebaseDatabase.getInstance().reference
     private val fiveThingsData = MutableLiveData<FiveThings>()
     private val fiveThingsDates = MutableLiveData<List<Date>>()
+    private val firebaseSource = FirebaseSource(user)
 
     fun getFiveThings(date: Date): LiveData<FiveThings> {
-        val formattedDate = getDatabaseStyleDate(date)
-
-        val dateQuery = database.child("users").child(user.uid).child(formattedDate)
-        Log.d("fivethings", "date query: " + dateQuery)
-        dateQuery.addValueEventListener(object : ValueEventListener {
-            override fun onCancelled(p0: DatabaseError?) {
-                Log.e("fivethings", p0.toString())
-            }
-
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                val results= dataSnapshot.value
-                if (results != null) {
-                    val things = results as ArrayList<String>
-                    val fiveThings = FiveThings(date,
-                            things[0],
-                            things[1],
-                            things[2],
-                            things[3],
-                            things[4],
-                            true)
-                    fiveThingsData.value = fiveThings
-                    Log.d("fivethings", "data set!")
-                } else {
-                    Log.d("fivethings", "no data found for this day")
-                    fiveThingsData.value = FiveThings(date, "", "","","","", false)
-                }
-            }
-        })
-        return fiveThingsData
+        return firebaseSource.getFiveThings(date)
     }
 
     fun onEditText() {
@@ -85,6 +58,16 @@ class FiveThingsViewModel(private val user: FirebaseUser): ViewModel() {
         }
     }
 
+
+//    fun writeFiveThings(fiveThings: FiveThings) {
+//        Log.d("fivethings", "about to write the data: " + fiveThings)
+//
+//
+//        firebaseSource.saveFiveThings(fiveThings, fiveThingsData)
+//
+//
+//    }
+
     fun getPreviousDay(date: Date): LiveData<FiveThings> {
         val prevDate = getPreviousDate(date)
         return getFiveThings(prevDate)
@@ -100,21 +83,7 @@ class FiveThingsViewModel(private val user: FirebaseUser): ViewModel() {
     }
 
     fun getWrittenDays(): LiveData<List<Date>> {
-        val query = database.child("users").child(user.uid)
-        query.addValueEventListener(object : ValueEventListener {
-            override fun onCancelled(p0: DatabaseError?) {
-                Log.e("fivethings", p0.toString())
-            }
-
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                val results= dataSnapshot.value as Map<String, List<String>>
-                val dayStrings = results.keys
-                Log.d("blerg", "dayStrings: " + dayStrings)
-                val days = dayStrings.map { getDateFromDatabaseStyle(it) }
-                fiveThingsDates.value = days
-            }
-        })
-        return fiveThingsDates
+        return firebaseSource.getWrittenDates()
     }
 
 }
